requirements.txt
mysql-connector-python
openpyxl
